[refactor] Extract question from Fragment
