+++ app/src/main/java/com/example/android/navigation/game/GameViewModel.kt
@@ -1,0 +1,8 @@
+package com.example.android.navigation.game
+
+import androidx.lifecycle.ViewModel
+
+class GameViewModel: ViewModel(){
+
+
+}
